精确番茄钟金币和修复切换场景时clock如果在计时就不会显示的问题，增加图片导出工具

+++ Assets/Scripts/New/Models/IAccountModel.cs
@@ -9,7 +9,7 @@ namespace BirdGame
         /// </summary>
         BindableProperty<float> Coins { get; }
 
-        int AddedCoins { get; set; }
+        float AddedCoins { get; set; }
     }
 
     public class AccountModel : AbstractModel, IAccountModel
@@ -19,6 +19,6 @@ namespace BirdGame
         }
 
         public BindableProperty<float> Coins { get; } = new BindableProperty<float>(600f);
-        public int AddedCoins { get; set; }
+        public float AddedCoins { get; set; }
     }
 }
